Handling the receiving and updating of movement states

+++ eden/gameserver/include/shaiya/game/world/model/Entity.hpp
@@ -74,7 +74,7 @@ namespace shaiya::game
          */
         bool hasUpdateFlag(UpdateMask mask) const
         {
-            return updateMask_ & mask;
+            return updateMask_ & static_cast<uint32_t>(mask);
         }
 
         /**
